+++ samples/java-tokenclient-usage/src/main/java/com/sap/cloud/security/xssec/samples/tokenflow/usage/HelloTokenClientServlet.java
@@ -1,0 +1,64 @@
+/**
+ * Copyright (c) 2019 SAP SE or an SAP affiliate company. All rights reserved.
+ * This file is licensed under the Apache Software License,
+ * v. 2 except as noted otherwise in the LICENSE file
+ * https://github.com/SAP/cloud-security-xsuaa-integration/blob/master/LICENSE
+ */
+package com.sap.cloud.security.xssec.samples.tokenflow.usage;
+
+import com.sap.cloud.security.xsuaa.client.ClientCredentials;
+import com.sap.cloud.security.xsuaa.client.DefaultOAuth2TokenService;
+import com.sap.cloud.security.xsuaa.client.OAuth2TokenResponse;
+import com.sap.cloud.security.xsuaa.client.XsuaaDefaultEndpoints;
+import com.sap.cloud.security.xsuaa.tokenflows.XsuaaTokenFlows;
+import org.json.JSONObject;
+
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+@WebServlet("/hello-token-client")
+public class HelloTokenClientServlet extends HttpServlet {
+	private static final long serialVersionUID = 1L;
+
+	/**
+	 * @see HttpServlet#doGet(HttpServletRequest request, HttpServletResponse
+	 * response)
+	 */
+	@Override
+	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
+		response.setContentType("text/plain");
+
+		JSONObject jsonObject = createJsonObjectFromVCAPServices();
+		String clientSecret = extractString(jsonObject, "/xsuaa/0/credentials/clientsecret");
+		String clientid = extractString(jsonObject, "/xsuaa/0/credentials/clientid");
+		String url = extractString(jsonObject, "/xsuaa/0/credentials/url");
+
+		XsuaaTokenFlows tokenFlows = new XsuaaTokenFlows(
+				new DefaultOAuth2TokenService(),
+				new XsuaaDefaultEndpoints(url), new ClientCredentials(clientid, clientSecret));
+		OAuth2TokenResponse tokenResponse = tokenFlows.clientCredentialsTokenFlow().execute();
+
+		writeLine(response, "Access-Token: " + tokenResponse.getAccessToken());
+		writeLine(response, "Access-Token-Payload: " + tokenResponse.getDecodedAccessToken().getPayload());
+		writeLine(response, "Expired-At: " + tokenResponse.getExpiredAtDate());
+
+	}
+
+	private String extractString(JSONObject jsonObject, String jsonPointer) {
+		return jsonObject.query(jsonPointer).toString();
+	}
+
+	private JSONObject createJsonObjectFromVCAPServices() {
+		String vcapServices = System.getenv("VCAP_SERVICES");
+		return new JSONObject(vcapServices);
+	}
+
+	private void writeLine(HttpServletResponse response, String string) throws IOException {
+		response.getWriter().append(string);
+		response.getWriter().append("\n");
+	}
+
+}