+++ token-client/src/main/java/com/sap/cloud/security/xsuaa/client/OAuth2TokenResponse.java
@@ -4,6 +4,7 @@ import com.sap.cloud.security.xsuaa.jwt.Base64JwtDecoder;
 import com.sap.cloud.security.xsuaa.jwt.DecodedJwt;
 
 import javax.annotation.Nullable;
+import java.time.Instant;
 import java.util.Date;
 import java.util.concurrent.TimeUnit;
 
@@ -42,8 +43,24 @@ public class OAuth2TokenResponse {
 		return new Base64JwtDecoder().decode(accessToken);
 	}
 
+	/**
+	 * Returns the moment in time when the token will be expired.
+	 *
+	 * @return the expiration point in time if present.
+	 * @deprecated use {@link #getExpiredAt()}.
+	 */
+	@Deprecated
 	public Date getExpiredAtDate() {
 		return new Date(expiredTimeMillis);
+	}
+
+	/**
+	 * Returns the moment in time when the token will be expired.
+	 *
+	 * @return the expiration point in time if present.
+	 */
+	public Instant getExpiredAt() {
+		return Instant.ofEpochMilli(expiredTimeMillis);
 	}
 
 	/**